argo-cd: Add volume parameters to Deployments
This adds new parameters to each Deployment in the argo-cd chart,
allowing users to specify custom volumes to mount.  This makes it
possible to use a Git repository self-signed certificates by manually
adding a "known_hosts" file, as recommended in the documentation:
https://argoproj.github.io/argo-cd/user-guide/private-repositories/

"volumes" and "volumeMounts" parameters are added to each Deployment,
with defaults set to empty.

This change was tested by templating the chart with default parameters,
and by templating it with volume mounts specified for each service.