Better Helm chart NOTES.txt templating
Add some extra templating to NOTES.txt, which are displayed to the
user after a successful helm install command.  These are really
helpful notes, but if you didn't specify a release name of argocd
or put it in a separate namespace, these commands weren't exactly
working with a copy paste.  Now they should take into account
the release name and release namespace helm was given.