Small CONTRIBUTING.md command line change
Instead of just doing `argocd version`, be a little more explicit
by passing in the server name and insecure flag.  This will make
sure that you're doing the version of the instance you're
port-forwarding above, as opposed to some other argocd.

While this works as is, I noted that before I ran the
`argocd login` command, I didn't have a default config with the
server name and insecure flag.  So the first time I did this,
the `argocd version` command gave an error:

FATA[0000] Argo CD server address unspecified