Better Helm NOTES.txt templating (#161)
* Better Helm chart NOTES.txt templating

Add some extra templating to NOTES.txt, which are displayed to the
user after a successful helm install command.  These are really
helpful notes, but if you didn't specify a release name of argocd
or put it in a separate namespace, these commands weren't exactly
working with a copy paste.  Now they should take into account
the release name and release namespace helm was given.

* Small CONTRIBUTING.md command line change

Instead of just doing `argocd version`, be a little more explicit
by passing in the server name and insecure flag.  This will make
sure that you're doing the version of the instance you're
port-forwarding above, as opposed to some other argocd.