feat: Argo CD allow custom paths for ingress
Similar to #317, this allows us to add extra paths to Argo CD.
This adds compatibility with the ALB Ingress Controller when using EKS
to be able to add the additional listener rule to perform a HTTP->HTTPS
redirect on port 80.

Unfortunately, we can't use the same structure as used in #317 because
there is an existing .Values.server.ingress.paths that we have to maintain
backwards compatibility with. The existing one is hard-coded to use a
fixed serviceName and servicePort, which doesn't work for the EKS ALB case.

Instead, we add .Values.server.ingress.extraPaths (and a similar var for
ingressGrpc) where a user directly specifies the entire rule, including
the path and backend information. Examples are added to the values.yaml
file for the extra path for EKS, with an empty list as a backwards
compatible default.