feat(argo-workflows) Include pod name in key format for GCS logs
A better default would be to add the `pod.name` to the key format of the Argo logs.

I noticed that when not including the pod name, it would constantly overwrite the same log from from different pods.
Including it makes sure that each pod will have its own log file.

Signed-off-by: Michiel De Mey <de.mey.michiel@gmail.com>