feat(argo-workflows): Add helper function to determine image value, minus tag
We have a situation where we want to exclude the registry so that images can
be pulled from in-network mirrors. Currently when you exclude the registry,
you get stuck with a prefixing forward slash.

Had to keep the tags in the manifest files because helm template requires
a colon in the image value.  Otherwise I'd move it into the helper function.

image entries changed: server, controller, executor

Signed-off-by: jmeridth <jmeridth@gmail.com>