fix(argo-workflows): Fix trailing whitespace in controller ConfigMap
This commit addresses two places where trailing whitespace may be
generated in the workflow-controller ConfigMap.

When the value `.resourceRateLimit` is not null, a trailing whitespace
is added after "resourceRateLimit:". This commit trims that whitespace.

When the value `.sso.redirectUrl` is left as the default of empty string,
a trailing whitespace is left after "redirectUrl:" as this value is not
quoted. This commit pipes this value through `quote` to ensure this
whitespace is no longer trailing and instead followed by `"` pair.

The main motivation of this commit is to generate a workflow-controller
ConfigMap that is properly pretty-printed, which trailing whitespace
prevents.

Signed-off-by: James Slater <SlaterByte@users.noreply.github.com>