fix(argo-workflows): revert `BASE_HREF` change (#2770)
* fix(argo-workflows): revert `BASE_HREF` change

This reverts commit 134cecde5812612514995205cd8fcc6a7f933373.

Sorry, I screwed up this and caused a regression. Didn't read my own release notes (from 4 months ago to be fair) where `ARGO_BASE_HREF` would not work previously as it was inconsistently `ARGO_BASEHREF`. So there is no forward compatible change to be made here, so just revert this.

Will have to reinstate this change for the 3.6 upgrade

Signed-off-by: Anton Gilgur <agilgur5@gmail.com>

* changelog update

Signed-off-by: Anton Gilgur <agilgur5@gmail.com>

---------

Signed-off-by: Anton Gilgur <agilgur5@gmail.com>